api-gateway/requirements.txt
Flask==2.3.3
PyJWT==2.8.0
requests==2.31.0
flasgger==0.9.7.1